architecture: canonicalize KDN service contracts
