+++ src/cacheroute/contracts/v1/__init__.py
@@ -1,4 +1,4 @@
-"""CacheRoute v1 common contract foundation."""
+"""Stable public CacheRoute v1 contract API."""
 
 from .common import (
     ContractModel, ENDPOINT_ID_PATTERN, GATEWAY_CONTRACT_VERSION,
@@ -6,10 +6,14 @@ from .common import (
     TokenReference, VersionedMessage, utc_now,
 )
 from .errors import ContractError, ContractErrorDetail, OutcomeCode
+from .knowledge import *
+from .knowledge import __all__ as _knowledge_all
+from .cache_service import *
+from .cache_service import __all__ as _cache_service_all
 
 __all__ = [
     "KDN_CONTRACT_VERSION", "GATEWAY_CONTRACT_VERSION", "ENDPOINT_ID_PATTERN",
     "SupportState", "utc_now", "ContractModel", "VersionedMessage",
     "GatewayTargetedRequest", "TokenReference", "TokenInput", "OutcomeCode",
     "ContractErrorDetail", "ContractError",
-]
+] + _knowledge_all + _cache_service_all
